backend/requirements.txt
Django>=5.0,<6.0
djangorestframework
djangorestframework-simplejwt
drf-spectacular
django-cors-headers
psycopg2-binary
gunicorn
django-environ
celery[redis]
redis
django-tenants
whitenoise
dj-database-url
pillow
djangorestframework-simplejwt[crypto]
google-auth>=2.23.0
